Modify vtxfmt code to always pass colors and indexes as floats. Modify vtxfmt code to include size 1,2,3 vertex attribute functions. Extend api_noop.c to include more functions for building an    outside-begin-end dispatch table or vtxfmt. Add opcodes to dlist.c to compile vertex attributes, begin/end commands, etc.    - The intention is for these to be used as a fallback for an optimizing 	display list compiler. Store Index internally as a float.

+++ src/mesa/main/mtypes.h
@@ -446,7 +446,7 @@ struct gl_current_attrib {
    /*@{*/
    GLfloat Attrib[VERT_ATTRIB_MAX][4];		/**< Current vertex attributes
 						  *  indexed by VERT_ATTRIB_* */
-   GLuint Index;				/**< Current color index */
+   GLfloat Index;				/**< Current color index */
    GLboolean EdgeFlag;				/**< Current edge flag */
    /*@}*/
